Add make_precompiler_downloader option
Allows users to customize download behavior, such as adding HTTP
authentication or using an alternate protocol like SFTP.
